// src/app/pages/models/sourcefunddata.model.ts
export interface sourcefundData {
            id: string;
            requestid: string;                
            qcontent: string; 
            rcontent: string; 
            qstatus: string; 
            subject: string; 
            agentid: string; 
       name: string;
       googleid: string;
       loginemail: string;
       loginphone: string;

       updatedate: string;
       country: string;
};